+++ Components/Individual/ImageCollage.tsx
@@ -1,8 +1,8 @@
 import { ImageList, ImageListItem } from "@mui/material";
-import banner1 from "../../assets/banner-long.png";
-import banner2 from "../../assets/banner2.png";
-import banner3 from "../../assets/banner3.png";
-import banner6 from "../../assets/banner6.png";
+import banner1 from "../../assets/banner-long.webp";
+import banner2 from "../../assets/banner2.webp";
+import banner3 from "../../assets/banner3.webp";
+import banner6 from "../../assets/banner6.webp";
 import React from "react";
 
 const imageData = [
