+++ package.json
@@ -15,7 +15,9 @@
     "@mui/material": "^5.8.1",
     "next": "12.1.6",
     "react": "18.1.0",
-    "react-dom": "18.1.0"
+    "react-dom": "18.1.0",
+    "react-virtualized-auto-sizer": "^1.0.6",
+    "react-window": "^1.8.7"
   },
   "devDependencies": {
     "@types/node": "17.0.35",
