+++ pages/index.tsx
@@ -1,60 +1,18 @@
 import { Card, CardContent, Grid, Stack, Typography } from "@mui/material";
-import banner from "../assets/banner-long.png";
-import banner2 from "../assets/banner2.png";
-import banner4 from "../assets/banner4.jpg";
-import banner5 from "../assets/banner5.png";
-import background from "../assets/background.png";
-import type { NextPage } from "next";
-import Head from "next/head";
+import CreativeCarousel from "../Components/Individual/CreativeCarousel";
 import ImageCollage from "../Components/Individual/ImageCollage";
+import type { NextPage } from "next";
 import { Box } from "@mui/system";
-import CreativeCarousel, {
-  CarouselProps,
-} from "../Components/Individual/CreativeCarousel";
+import Head from "next/head";
 
 /**
  * @fileoverview Website homepage
  * @returns NextPage - A Nextjs Page
  */
-
-//Data for populating carousel
-const carouselData = [
-  {
-    src: banner,
-    heading: "Author",
-    body: "Bosede is a published author and continues to write for the growth of others but also for herself. Her subjects range from spirituality and memoirs to self-help and empowerment. You can gleam some of her work in the bookstore here",
-    link: "/bookstore"
-  },
-  {
-    src: banner4,
-    heading: "Trailblazer",
-    body: "This esteemed woman leads by example. This means marching with those she wants to protect and proving that anytihng is possible with enough passion and dedication!",
-    link: "/esteemed"
-  },
-  {
-    src: background,
-    heading: "Motivational Speaker",
-    body: "Through talks in person and online in the US and to Esteemed Woman groups across the globe, Bosede lifts people up and motivates them to chase their ideas and goals. Check out her outlets here!",
-    link: "/esteemed"
-  },
-  {
-    src: banner5,
-    heading: "Leadership Coach",
-    body: "She was coached and certified under the leadership of John C. Maxwell. Bosede equips individuals and organizations with practical tools to break barriers, maximize their strengths, and amplify their success throughout all walks of life.",
-    link: "leadership"
-  },
-  {
-    src: banner2,
-    heading: "Licensed CPR Instructor",
-    body: "Bosede is also a licensed CPR instructor, offering classes to groups and individuals looking for a certification. Whether its required for work or if you just want the skill, schedule a class today!",
-    link: "/cpr"
-  },
-];
 const Home: NextPage = () => {
   return (
     <>
       <Head>
-        {" "}
         {/*Metadate of the page*/}
         <title>Home - Bosede Adetunji</title>
         <meta
@@ -115,10 +73,11 @@ const Home: NextPage = () => {
           I am a _______
         </Typography>
 
-        <Box
+        <Grid
           component={Stack}
           direction="row"
           spacing={5}
+          xs={12}
           sx={{
             overflowX: "scroll",
             overflowY: "hidden",
@@ -126,16 +85,8 @@ const Home: NextPage = () => {
             p: [0, 4, 4, 4],
           }}
         >
-          {carouselData.map((item: CarouselProps) => (
-            <CreativeCarousel
-            key={item.heading}
-            src={item.src}
-            heading={item.heading}
-            body={item.body}
-            link={item.link}
-          />
-          ))}
-        </Box>
+          <CreativeCarousel />
+        </Grid>
       </main>
     </>
   );
